Dockerfile: Rename syntax file to correct case
The dockerfile syntax definition was erroniously named "dockerfile"
instead of "Dockerfile". This caused vim to be unable to find the syntax
highlighting for Dockerfiles because the ftdetect logic for the
ekalinin/Dockerfile.vim plugin sets ft to 'Dockerfile'.

Running the 'build' script on a platform with a case sensitive file
system correctly copies the file, but on a system that is case
insensitive and case preserving, the file name would not change the
build system is not written to clean up existing files before building.

Manually renaming the file in git will prevent the issue from recuring
in future builds without needing to run an update for address a file
system change.