Check if Idris version is specified at start of file.
Added two heuristics which check if the user has specified the version
of Idris used in a comment at the start of the file, e.g.
`-- Using Idris 2`. As far as I can tell, this works correctly: the
`idris.vim` plugin is loaded when specifying Idris 1 at the start, and
the `idris2.vim` files are loaded when specifying Idris 2.