Added heuristics for literate Idris 1 and 2.
Literate Idris 1 and 2 pose a bit of a problem because Idris's comment
denotator `--` could be interpreted differently in plaintext/literate
programming. Currently, the heuristics assume there will be a
declaration of the version of Idris in the first code block used, but
this could probably be improved, e.g. by having a preamble or something
similar.