Update the declaration of ft.vim
According to how I read the `vim9script` documentation, autoload
functions don't need to be exported but still need to be prefixed by the
autoloaded path and filename.

> When using `:function` or `:def` to specify a new function at the
> script level in a Vim9 script, the function is local to the script, as
> if "s:" was prefixed.  Using the "s:" prefix is optional.  To define
> a global function or variable the "g:" prefix must be used.  For
> functions in an autoload script the "name#" prefix is sufficient.

>         def ThisFunction()          # script-local
>         def s:ThisFunction()        # script-local
>         def g:ThatFunction()        # global
>         def scriptname#function()   # autoload

The way that I read this is that `export def scriptname#function()` is
not required, because `def scriptname#function()` means the same thing.
What is not clear — except that _not_ including `polyglot#ft#` is
breaking various Read events from firing — is whether `export def Foo`
causes the inheritance of the `name#` prefix. It appears not to be the
case by default.